+++ helen/__init__.py
@@ -1,3 +1,3 @@
 """helen package."""
 
-__version__ = "1.39.8"
+__version__ = "1.39.9"